+++ h5netcdf/core.py
@@ -1184,7 +1184,8 @@ class File(Group):
         Parameters
         ----------
         path: path-like
-            Location of the netCDF4 file to be accessed.
+            Location of the netCDF4 file to be accessed, or an h5py File object,
+            or a Python file-like object (which should read/write bytes).
 
         mode: "r", "r+", "a", "w"
             A valid file access mode. Defaults to "r".
@@ -1219,6 +1220,10 @@ class File(Group):
 
         Datasets created with h5netcdf version 0.12.0 that are opened with
         newer versions of h5netcdf will continue to disable order tracker.
+
+        If an h5py File object is passed in, closing the h5netcdf wrapper will
+        not close the h5py File. In other cases, closing the h5netcdf File object
+        does close the underlying file.
         """
         # 2022/01/09
         # netCDF4 wants the track_order parameter to be true
@@ -1236,6 +1241,7 @@ class File(Group):
         track_order = kwargs.pop("track_order", track_order_default)
 
         self.decode_vlen_strings = kwargs.pop("decode_vlen_strings", None)
+        self._close_h5file = True
         try:
             if isinstance(path, str):
                 if (
@@ -1263,6 +1269,12 @@ class File(Group):
                     self._h5file = self._h5py.File(
                         path, mode, track_order=track_order, **kwargs
                     )
+            elif isinstance(path, h5py.File):
+                self._preexisting_file = mode in {"r", "r+", "a"}
+                self._h5py = h5py
+                self._h5file = path
+                # h5py File passed in: let the caller decide when to close it
+                self._close_h5file = False
             else:  # file-like object
                 self._preexisting_file = mode in {"r", "r+", "a"}
                 self._h5py = h5py
@@ -1402,7 +1414,9 @@ class File(Group):
     def close(self):
         if not self._closed:
             self.flush()
-            self._h5file.close()
+            if self._close_h5file:
+                self._h5file.close()
+            self._h5file = None
             self._closed = True
 
     __del__ = close